+++ tests/Feature/HealthTest.php
@@ -26,6 +26,7 @@ class HealthTest extends TestCase
                     'name',
                     'php',
                     'laravel',
+                    'butlerService',
                     'timezone',
                 ],
                 'checks' => [